move components to dedicated folder
