+++ client/src/components/onboarding/ProgressBar.js
